Refactor: SequenceId to ushort, add collection tests/docs
- Change all packet SequenceId fields from uint (32-bit) to ushort (16-bit) across interfaces, classes, header offsets, and serialization logic
- Update PacketHeaderOffset and header extension methods for new SequenceId size
- Add comprehensive tests for serialization/deserialization of complex and nested collections, including null/empty, Unicode, large payloads, and special values
- Optimize type metadata warmup to skip internal fields of standard collections
- Add enums.md documenting all system-wide enumerations; update index
- Minor doc and test value improvements; add copyright headers

+++ src/Nalix.Common/Networking/Packets/IPacket.cs
@@ -60,7 +60,7 @@ public interface IPacket
     /// Gets the sequence identifier used to correlate requests and responses.
     /// </summary>
     [SerializeHeader(PacketHeaderOffset.SequenceId)]
-    uint SequenceId { get; }
+    ushort SequenceId { get; }
 
     #endregion Metadata
